+++ src/main/java/Console.java
@@ -3,13 +3,14 @@ import java.util.Scanner;
 
 public class Console {
 
-
     UserWarehouse UW = new UserWarehouse();
     HashMap<String, Integer> userList = UW.getAccounts();
     HashMap<String, User> objList = UW.getObjectList();
     StringBuilder transactionHistory = new StringBuilder();
     Menu menu = new Menu();
-
+    CheckingMenu checkingMenu = new CheckingMenu();
+    SavingsMenu savingsMenu = new SavingsMenu();
+    InvestmentMenu investmentMenu = new InvestmentMenu();
 
     public Console() {
     }
@@ -75,21 +76,36 @@ public class Console {
             switch (selection) {
                 case 1:
                     System.out.println("Checking");
-                    checkAcctMenu(current);
+                     if(current.getHasChecking()){
+                    checkingMenu.checkAcctMenu(current);}
+                     else{CheckingMenu.verifyAccount(current);
+                         if(current.getHasChecking()){
+                             checkingMenu.checkAcctMenu(current);}}
                     break;
                 case 2:
                     System.out.println("Savings");
-                    savingsAcctMenu(current);
+                    if(current.getHasSavings()){
+                        savingsMenu.savingsAcctMenu(current);}
+                    else{
+                        SavingsMenu.verifyAccount(current);
+                        if(current.getHasSavings()){
+                            savingsMenu.savingsAcctMenu(current);
+                        }
+                    }
                     break;
                 case 3:
                     System.out.println("Investment");
-                    investmentAcctMenu(current);
+                    if(current.getHasInvestment()){
+                        investmentMenu.investmentAcctMenu(current);
+                    }
+                    else{
+                        InvestmentMenu.verifyAccount(current);
+                        if(current.getHasInvestment()){
+                            investmentMenu.investmentAcctMenu(current);
+                        }
+                    }
                     break;
                 case 4:
-                    System.out.println("Transaction History");
-                    getTransactions();
-                    break;
-                case 5:
                     quit = true;
                     break;
                 default:
@@ -100,256 +116,5 @@ public class Console {
         verifyUser();
     }
 
-    public void getTransactions() {
-        String trans = transactionHistory.toString();
-        System.out.print(trans);
-    }
-
-    public void checkAcctMenu(User current) {
-        boolean checking = current.getHasChecking();
-        if (checking == true) {
-            boolean back = false;
-            while (back == false) {
-                menu.acctMenu();
-                Scanner option = new Scanner(System.in);
-                int selection = option.nextInt();
-                switch (selection) {
-                    case 1:
-                        System.out.println("Enter amount to withdraw");
-                        Scanner withD = new Scanner(System.in);
-                        double withdraw = withD.nextDouble();
-                        System.out.println("You withdrew: ");
-                        System.out.println(current.checkingWithdraw(withdraw));
-                        System.out.println("current balance: ");
-                        System.out.println(current.getCheckingBalance());
-                        transactionHistory.append("Withdrew: " + withdraw + " from checking. " + "Remaining Balance: " + current.getCheckingBalance() + "\n");
-                        break;
-                    case 2:
-                        System.out.println("Enter amount to deposit");
-                        Scanner depos = new Scanner(System.in);
-                        double deposit = depos.nextDouble();
-                        System.out.println("You deposited: ");
-                        current.checkingDeposit(deposit);
-                        System.out.println("current balance: ");
-                        System.out.println(current.getCheckingBalance());
-                        transactionHistory.append("Deposited: " + deposit + " into checking. " + "Remaining Balance: " + current.getCheckingBalance() + "\n");
-                        break;
-                    case 3:
-                        System.out.println(current.getCheckingBalance());
-                        break;
-                    case 4:
-                        System.out.println("Enter account to transfer to");
-                        Scanner type = new Scanner(System.in);
-                        String acct = type.nextLine();
-                        System.out.println("Enter amount to transfer");
-                        Scanner amt = new Scanner(System.in);
-                        double amount = amt.nextDouble();
-                        if (acct.toLowerCase().equals("savings") && current.getHasSavings()==true) {
-                            current.checkingWithdraw(amount);
-                            current.savingsDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: " + current.getCheckingBalance());
-                        } else if (acct.toLowerCase().equals("investment") && current.getHasInvestment()==true) {
-                            current.checkingWithdraw(amount);
-                            current.investmentDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: " + current.getCheckingBalance());
-                        } else {
-                            System.out.println("Account not found. Please create account before transfer.");
-                        }
-                        break;
-                    case 5:
-                        current.setHasChecking(false);
-                        System.out.println(current.closeCheck());
-                        accountSelectionMenu(current);
-                        break;
-                    case 6:
-                        back = true;
-                        break;
-                    default:
-                        System.out.println("Incorrect input");
-                        break;
-
-                }
-            }
-        } else {
-            System.out.println("Would you like to create a checking account?\n" +
-                    "Enter yes or no");
-            Scanner scan = new Scanner(System.in);
-            String answer = scan.nextLine();
-            if (answer.toLowerCase().equals("yes")) {
-                current.setHasChecking(true);
-                System.out.println("Account creation successful.\nChecking:");
-
-                checkAcctMenu(current);
-            } else {
-                accountSelectionMenu(current);
-            }
-        }
-    }
-
-
-    public void savingsAcctMenu(User current) {
-        boolean savings = current.getHasSavings();
-        if (savings == true) {
-            boolean back = false;
-            while (back == false) {
-                menu.acctMenu();
-                Scanner option = new Scanner(System.in);
-                int selection = option.nextInt();
-                switch (selection) {
-                    case 1:
-                        System.out.println("Enter amount to withdraw");
-                        Scanner withD = new Scanner(System.in);
-                        double withdraw = withD.nextDouble();
-                        System.out.println("You withdrew: ");
-                        System.out.println(current.savingsWithdraw(withdraw));
-                        System.out.println("current balance: ");
-                        System.out.println(current.getSavingBalance());
-                        transactionHistory.append("Withdrew: " + withdraw + " from savings. " + "Remaining Balance: " + current.getSavingBalance() + "\n");
-                        break;
-                    case 2:
-                        System.out.println("Enter amount to deposit");
-                        Scanner depos = new Scanner(System.in);
-                        double deposit = depos.nextDouble();
-                        System.out.println("You deposited: ");
-                        current.savingsDeposit(deposit);
-                        System.out.println("current balance: ");
-                        System.out.println(current.getSavingBalance());
-                        transactionHistory.append("Deposited: " + deposit + " into savings. " + "Remaining Balance: " + current.getSavingBalance() + "\n");
-                        break;
-                    case 3:
-                        System.out.println(current.getSavingBalance());
-                        break;
-                    case 4:
-                        System.out.println("Enter account to transfer to");
-                        Scanner type = new Scanner(System.in);
-                        String acct = type.nextLine();
-                        System.out.println("Enter amount to transfer");
-                        Scanner amt = new Scanner(System.in);
-                        double amount = amt.nextDouble();
-                        if (acct.toLowerCase().equals("checking") && current.getHasChecking()==true) {
-                            current.savingsWithdraw(amount);
-                            current.checkingDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: " + current.getCheckingBalance());
-                        } else if (acct.toLowerCase().equals("investment") && current.getHasInvestment()==true) {
-                            current.savingsWithdraw(amount);
-                            current.investmentDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: " + current.getCheckingBalance());
-                        } else {
-                            System.out.println("Account not found. Please create account before transfer.");
-                        }
-                        break;
-                    case 5:
-                        current.setHasSavings(false);
-                        System.out.println(current.closeSaving());
-                        accountSelectionMenu(current);
-                        break;
-                    case 6:
-                        back = true;
-                        break;
-                    default:
-                        System.out.println("Incorrect input");
-                        break;
-
-                }
-
-
-            }
-
-        } else {
-            System.out.println("Would you like to create a savings account?\n" +
-                    "Enter yes or no");
-            Scanner scan = new Scanner(System.in);
-            String answer = scan.nextLine();
-            if (answer.toLowerCase().equals("yes")) {
-                current.setHasSavings(true);
-                System.out.println("Account creation successful.\nSavings:");
-                savingsAcctMenu(current);
-            } else {
-                accountSelectionMenu(current);
-            }
-        }
-    }
-
-    public void investmentAcctMenu(User current) {
-        boolean investment = current.getHasInvestment();
-        if (investment == true) {
-            boolean back = false;
-            while (back == false) {
-                menu.acctMenu();
-                Scanner option = new Scanner(System.in);
-                int selection = option.nextInt();
-                switch (selection) {
-                    case 1:
-                        System.out.println("Enter amount to withdraw");
-                        Scanner withD = new Scanner(System.in);
-                        double withdraw = withD.nextDouble();
-                        System.out.println("You withdrew: ");
-                        System.out.println(current.investmentWithdraw(withdraw));
-                        System.out.println("current balance: ");
-                        System.out.println(current.getInvestmentBalance());
-                        transactionHistory.append("Withdrew: " + withdraw + " from investment. " + "Remaining Balance: " + current.getInvestmentBalance() + "\n");
-                        break;
-                    case 2:
-                        System.out.println("Enter amount to deposit");
-                        Scanner depos = new Scanner(System.in);
-                        double deposit = depos.nextDouble();
-                        System.out.println("You deposited: ");
-                        current.investmentDeposit(deposit);
-                        System.out.println("current balance: ");
-                        System.out.println(current.getInvestmentBalance());
-                        transactionHistory.append("Deposited: " + deposit + " into investment. " + "Remaining Balance: " + current.getInvestmentBalance() + "\n");
-                        break;
-                    case 3:
-                        System.out.println(current.getInvestmentBalance());
-                        break;
-                    case 4:
-                        System.out.println("Enter account to transfer to");
-                        Scanner type = new Scanner(System.in);
-                        String acct = type.nextLine();
-                        System.out.println("Enter amount to transfer");
-                        Scanner amt = new Scanner(System.in);
-                        double amount = amt.nextDouble();
-                        if(acct.toLowerCase().equals("checking") && current.getHasChecking()==true){
-                            current.investmentWithdraw(amount);
-                            current.checkingDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: "+current.getCheckingBalance());
-                        }
-                        else if(acct.toLowerCase().equals("savings")&& current.getHasSavings()==true){
-                            current.investmentWithdraw(amount);
-                            current.savingsDeposit(amount);
-                            System.out.println("Transfer successful \nCurrent balance: "+current.getCheckingBalance());
-                        }
-                        else{
-                            System.out.println("Account not found. Please create account before transfer.");
-                        }
-                        break;
-                    case 5:
-                        current.setHasInvestment(false);
-                        System.out.println(current.closeInvestment());
-                        accountSelectionMenu(current);
-                        break;
-                    case 6:
-                        back = true;
-                        break;
-                    default:
-                        System.out.println("Incorrect input");
-                        break;
-
-                }
-            }
-        } else {
-            System.out.println("Would you like to create an investment account?\n" +
-                    "Enter yes or no");
-            Scanner scan = new Scanner(System.in);
-            String answer = scan.nextLine();
-            if (answer.toLowerCase().equals("yes")) {
-                current.setHasInvestment(true);
-                System.out.println("Account creation successful.\nInvestment:");
-                investmentAcctMenu(current);
-            } else {
-                accountSelectionMenu(current);
-            }
-        }
-    }
 }
